* implemented: advice/session history

+++ data/src/main/java/com/nikasov/data/local/entity/AdviceEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "advices")
 data class AdviceEntity(
     @PrimaryKey(autoGenerate = true)
-    val adviceId: Int = 0,
-    val sessionId: Int,
+    val adviceId: Long = 0,
+    val sessionId: Long,
     val text: String
 )
